skip push of old files
The sure way to remove a file is to remove it from the local upsync cache and also "upspin rm" it.  But that is tedious if you're also working elsewhere with upspinfs.

So we try to set and use the atime of the local upsync root as a hint of the last time upsync was completed.  If a file in the local cache is older than that but not in the upspin master copy, most likely it was deleted via upspinfs.  So skip pushing it, and print a warning.