record upsync time in ~/upspin/upsync
The heuristic for "skipping old" needs a place to store the last upsync time.  Modifying the time of the local sync root doesn't work because that gets updated by other local file operations.  Instead, let's create yet another file in ~/upspin.

That also has the side benefit of providing a reminder of which local directory we're using for upsync.